+++ Sources/BraintreeCore/Analytics/BTAnalyticsServiceError.swift
@@ -1,20 +1,20 @@
 import Foundation
 
 ///  Error codes associated with a API Client.
-enum BTAnalyticsServiceError: Int, Error, CustomNSError, LocalizedError {
+public enum BTAnalyticsServiceError: Int, Error, CustomNSError, LocalizedError, Equatable {
 
     /// 0. Invalid API client
     case invalidAPIClient
 
-    static var errorDomain: String {
+    public static var errorDomain: String {
         "com.braintreepayments.BTAnalyticsServiceErrorDomain"
     }
 
-    var errorCode: Int {
+    public var errorCode: Int {
         rawValue
     }
 
-    var errorDescription: String? {
+    public var errorDescription: String? {
         switch self {
         case .invalidAPIClient:
             return "API client must have client token or tokenization key"