+++ Sources/BraintreeDataCollector/BTDataCollectorError.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 ///  Error details associated with Braintree Data Collector.
-enum BTDataCollectorError: Int, Error, CustomNSError, LocalizedError {
+public enum BTDataCollectorError: Int, Error, CustomNSError, LocalizedError, Equatable {
 
     /// 0. Unknown error
     case unknown
@@ -12,15 +12,15 @@ enum BTDataCollectorError: Int, Error, CustomNSError, LocalizedError {
     /// 2. The device data could not be encoded.
     case encodingFailure
 
-    static var errorDomain: String {
+    public static var errorDomain: String {
         "com.braintreepayments.BTDataCollectorErrorDomain"
     }
     
-    var errorCode: Int {
+    public var errorCode: Int {
         rawValue
     }
 
-    var errorDescription: String? {
+    public var errorDescription: String? {
         switch self {
         case .unknown:
             return "An unknown error occurred. Please contact support."